+++ web-app/package.json
@@ -18,7 +18,6 @@
         "@mantine/dropzone": "^8.3.16",
         "@mantine/form": "^8.3.16",
         "@mantine/hooks": "^8.3.16",
-        "@react-pdf/renderer": "^4.3.2",
         "@spotify/web-api-ts-sdk": "^1.2.0",
         "@supabase/supabase-js": "^2.97.0",
         "@tabler/icons-react": "^3.37.1",
