+++ web-app/src/features/pdf/pdf-lib/constants.ts
@@ -1,0 +1,18 @@
+export const A4_WIDTH = 595.28;
+export const A4_HEIGHT = 841.89;
+export const PAGE_PADDING = 30;
+
+export const CARD_WIDTH = 170;
+export const CARD_HEIGHT = 170;
+
+export const CARDS_PER_ROW_DOUBLE = 3;
+export const ROWS_PER_PAGE_DOUBLE = 4;
+export const CARDS_PER_PAGE_DOUBLE = CARDS_PER_ROW_DOUBLE * ROWS_PER_PAGE_DOUBLE;
+export const DOUBLE_SIDED_COLUMN_GAP = 0;
+export const DOUBLE_SIDED_ROW_GAP = 0;
+
+export const CARDS_PER_ROW_ONE = 3;
+export const ROWS_PER_PAGE_ONE = 2;
+export const CARDS_PER_PAGE_ONE = CARDS_PER_ROW_ONE * ROWS_PER_PAGE_ONE;
+export const ONE_SIDED_COLUMN_GAP = 8;
+export const ONE_SIDED_ROW_GAP = 24;
